Move test base into more generic name space
It is usable for testing more than blocks.

Also move generation of mock methods into test cases where they are
needed. We register mock functions with namespace and this will differ
between tests.
