+++ web/modules/custom/dbcdk_community_moderation/tests/src/Unit/Plugin/Block/ProfilesBlockTest.php
@@ -11,14 +11,31 @@ use DBCDK\CommunityServices\Model\Profile as ModelProfile;
 use DBCDK\CommunityServices\Model\Quarantine;
 use Drupal\dbcdk_community_moderation\Plugin\Block\ProfilesBlock;
 use Drupal\dbcdk_community_moderation\Profile\Profile;
-use Drupal\Tests\dbcdk_community\Unit\Plugin\Block\BlockTestBase;
+use Drupal\Tests\dbcdk_community\Unit\UnitTestBase;
 
 /**
  * Test class for Profiles block.
  *
  * @group dbcdk_community_moderation
  */
-class ProfilesBlockTest extends BlockTestBase {
+class ProfilesBlockTest extends UnitTestBase {
+
+  /* @var \PHPUnit_Framework_MockObject_MockObject */
+  protected $pager;
+
+  /* @var \PHPUnit_Framework_MockObject_MockObject */
+  protected $message;
+
+  /**
+   * {@inheritdoc}
+   */
+  public function setUp() {
+    parent::setUp();
+
+    $namespace = '\Drupal\dbcdk_community_moderation\Plugin\Block';
+    $this->pager = $this->getFunctionMock($namespace, 'pager_default_initialize');
+    $this->message = $this->getFunctionMock($namespace, 'drupal_set_message');
+  }
 
   /**
    * Generate a new ProfilesBlock wired up with mocks.
